Move files to separate library folders
