+++ mirrulations-dashboard/src/static/index.js
