+++ mirrulations-dashboard/src/static/style.css
