+++ mirrulations-dashboard/src/templates/index.html
