+++ mirrulations-work-generator/setup.cfg
@@ -1,0 +1,28 @@
+[metadata]
+name = capstone2021
+
+[options]
+package_dir=
+    =src
+packages=find:
+install_requires =
+    fakeredis >= 1.4.5
+    flake8 >= 3.8.4
+    Flask >= 1.1.2
+    pycodestyle >= 2.6.0
+    pylint >= 2.7.2
+    pytest >= 6.2.2
+    pytest-cov >= 2.11.1
+    redis >= 3.5.3
+    requests >= 2.25.1
+    fakeredis >= 1.4.5
+    requests-mock >= 1.8.0
+    pytest-mock >= 3.5.1
+    Flask-Cors >= 3.0.10
+    python-dotenv >= 0.17.0
+    wheel >= 0.36.2
+    gunicorn >= 20.1.0
+    pytz >= 2021.1
+    pymongo >= 3.11.4
+[options.packages.find]
+where=src
