+++ mirrulations-work-server/test_work_server.py
@@ -2,7 +2,16 @@ from json import dumps
 from fakeredis import FakeRedis, FakeServer
 from pytest import fixture
 from c21server.work_server.work_server import create_server
-from .test_utils import mock_flask_server
+
+
+def mock_flask_server(create_server):
+    redis_server = FakeServer()
+    mock_db = FakeRedis(server=redis_server)
+    server = create_server(mock_db)
+    server.redis_server = redis_server
+    server.app.config['TESTING'] = True
+    server.client = server.app.test_client()
+    return server
 
 
 @fixture(name='mock_server')
